refactor: optimise dynaimc favicon
Old method involved manually painting the favicons and generating PNG dataURIs at runtime. This method has the SCG dataURIs cached in the constants file, so now we just switch them out when appropriate.
Maintains the benefit of only needing this code when the sessino page loads (future-proofing for code splitting)